task

+++ app/Http/Controllers/TasksController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use Illuminate\Support\Facades\Auth;  
 use App\Models\Task;
 
 class TasksController extends Controller
@@ -16,11 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::orderBy('id', 'desc')->paginate(10);
+        $data = [];
+        // 認証済みの場合
+        if(\Auth::check()){ 
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // タスク一覧を作成日時の降順で取得(自身のタスクのみ取得)
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
 
     /**
@@ -48,12 +57,20 @@ class TasksController extends Controller
         $request->validate([
             'status' => 'required|max:10',
             'content' => 'required',
-            ]);
-            
+        ]);
+        
+         /*
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        */
+        
+        // 認証済みユーザのタスクとして作成
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
         
         return redirect('/');
     }
@@ -66,11 +83,20 @@ class TasksController extends Controller
      */
    public function show($id)
     {
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザの場合、タスク詳細ビューを表示
+        if (\Auth::id() === $task->user_id) {
+            
+            return view("tasks.show", [
+                "task" => $task,
+            ]);
+        }
+        
+        // トップページにリダイレクトさせる
+        return redirect("/")
+            ->with("failed");
     }
 
     /**
@@ -81,11 +107,20 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        return view('tasks.edit',[
-            'task' => $task,
-        ]);
+        // 認証済みユーザの場合、タスク編集ビューを表示
+        if (\Auth::id() === $task->user_id) {
+            
+            return view('tasks.edit',[
+                'task' => $task,
+            ]);
+        }
+        
+        // トップページにリダイレクトさせる
+        return redirect("/")
+            ->with("failed");
     }
 
     /**
@@ -101,7 +136,8 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required',
             ]);
-            
+        
+        // idの値でタスクを検索して取得    
         $task = Task::findOrFail($id);
         
         $task->status = $request->status;
@@ -119,9 +155,18 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        $task->delete();
         
-        return redirect('/');
+        //認証済みユーザがその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect("/")
+                ->with('success','Delete Successful');
+        }
+
+        // トップページにリダイレクトさせる
+        return redirect("/")
+            ->with('Delete Failed');
     }
 }
